added safety
- added safety to delete url and set active
- if user try to access profile without being connected he get redirected
- added Bdd.validateUrlOwner method
- changed SQL request of getUserUrls and validateUrlOwner to join instead of doing two SQL request
- update database diagram users.email should now be UNIQUE
- added some comments

+++ controllers/profile_controller.php
@@ -3,11 +3,18 @@
 if (userConnected()) {
     $urls = $bdd->getUserUrls($_SESSION['user']);
 }
+else{
+    header("Location: index.php?page=login");
+}
 
 // Delete
 if (isset($_GET['delete'])) {
     $id = filter_input(INPUT_GET, 'delete');
-    // TODO: check if the user is the owner of the url
+
+    if(!$bdd->validateUrlOwner($id, $_SESSION['user'])){
+        header("Location: index.php?page=profile");
+        die();
+    }
 
     $bdd->deleteUrlByID($id);
     unset($_GET['delete']);
@@ -18,7 +25,11 @@ if (isset($_GET['delete'])) {
 if (isset($_GET['updateId']) && isset($_GET['updateActive'])) {
     $id = filter_input(INPUT_GET, 'updateId');
     $is_active = filter_input(INPUT_GET, 'updateActive');
-    // TODO: check if the user is the owner of the url
+
+    if(!$bdd->validateUrlOwner($id, $_SESSION['user'])){
+        header("Location: index.php?page=profile");
+        die();
+    }
 
     $bdd->setActive($id, $is_active ? 0 : 1);
     unset($_GET['updateId']);
